blazev2 and CLEAN

+++ Assets/Effects/FlashShiftShaderData.cs
@@ -6,38 +6,38 @@ using Terraria.Graphics.Shaders;
 
 namespace MetroidMod.Assets.Effecs
 {
-    public class FlashShiftShaderData : ArmorShaderData
-    {
+	public class FlashShiftShaderData : ArmorShaderData
+	{
 
-        private static bool isInitialized;
+		private static bool isInitialized;
 
-        private static ArmorShaderData dustShaderData;
+		private static ArmorShaderData dustShaderData;
 
-        public FlashShiftShaderData(Ref<Effect> shader, string passName) : base(shader, passName)
-        {
-            if (!isInitialized)
-            {
-                isInitialized = true;
+		public FlashShiftShaderData(Ref<Effect> shader, string passName) : base(shader, passName)
+		{
+			if (!isInitialized)
+			{
+				isInitialized = true;
 				dustShaderData = new ArmorShaderData(shader, passName).UseOpacity(1);
-            }
-        }
+			}
+		}
 		public override void Apply(Entity entity, DrawData? drawData)
-        {
-            Player player = entity as Player;
-            if (entity != null && entity is Projectile)
-            {
-                Projectile p = entity as Projectile;
-                if (p != null) 
-                    player = Main.player[p.owner];
-            }
-            if (player == null)
-            {
-                return;
-            }
+		{
+			Player player = entity as Player;
+			if (entity != null && entity is Projectile)
+			{
+				Projectile p = entity as Projectile;
+				if (p != null)
+					player = Main.player[p.owner];
+			}
+			if (player == null)
+			{
+				return;
+			}
 			float o = 1f;
-			o = (float)player.GetModPlayer<MPlayer>().flashShiftTime / (float)player.GetModPlayer<MPlayer>().flashShiftLength;
+			o = player.GetModPlayer<MPlayer>().flashShiftTime / (float)player.GetModPlayer<MPlayer>().flashShiftLength;
 
-			float s = o; 
+			float s = o;
 			if (o <= 0.2 && player.GetModPlayer<MPlayer>().flashShiftGlow)
 			{
 				o = 0.2f;
@@ -46,8 +46,8 @@ namespace MetroidMod.Assets.Effecs
 			base.UseColor(player.GetModPlayer<MPlayer>().flashShiftColor);
 			base.UseOpacity(o);
 			base.UseSaturation(s);
-            base.Apply(player, drawData);
-        }
+			base.Apply(player, drawData);
+		}
 
 		public override ArmorShaderData GetSecondaryShader(Entity entity)
 		{
